refactor(#3): Minor refactor
