+++ user/src/main/java/org/ubdev/user/dto/JoinTripRequestDataDto.java
@@ -1,10 +1,10 @@
-package org.ubdev.user.model;
+package org.ubdev.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.util.UUID;
 
-public record JoinTripRequestData (
+public record JoinTripRequestDataDto(
         @JsonProperty("user_id")
         UUID userId,
         String name,
