+++ user/src/main/java/org/ubdev/user/dto/UserUpdateDto.java
@@ -1,14 +1,11 @@
 package org.ubdev.user.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
-import org.springframework.web.multipart.MultipartFile;
 import org.ubdev.user.model.Preference;
 
 public record UserUpdateDto (
         String name,
         String surname,
-
-        MultipartFile photo,
 
         @JsonProperty("phone_number")
         String phoneNumber,
