+++ user/src/main/java/org/ubdev/user/service/UserServiceImpl.java
@@ -5,6 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.multipart.MultipartFile;
 import org.ubdev.kafka.model.user.CreateUserMessage;
 import org.ubdev.kafka.model.user.DeleteUserMessage;
 import org.ubdev.kafka.model.user.UpdateEmailMessage;
@@ -12,7 +13,7 @@ import org.ubdev.user.dto.UserDto;
 import org.ubdev.user.dto.UserUpdateDto;
 import org.ubdev.user.exception.exceptions.UserNotFoundException;
 import org.ubdev.user.mapper.UserMapper;
-import org.ubdev.user.model.JoinTripRequestData;
+import org.ubdev.user.dto.JoinTripRequestDataDto;
 import org.ubdev.user.model.User;
 import org.ubdev.user.repository.UserRepository;
 import org.ubdev.util.cache.UserCacheService;
@@ -52,11 +53,11 @@ public class UserServiceImpl implements UserService {
 
     @Transactional(readOnly = true)
     @Override
-    public JoinTripRequestData getJoinTripRequestData(String email, UUID ownerId) {
+    public JoinTripRequestDataDto getJoinTripRequestData(String email, UUID ownerId) {
         User user = cacheService.getUserById(ownerId);
         String ownerEmail = userRepository.findEmailById(ownerId);
 
-        return new JoinTripRequestData(user.getId(), user.getName(), user.getSurname(), ownerEmail);
+        return new JoinTripRequestDataDto(user.getId(), user.getName(), user.getSurname(), ownerEmail);
     }
 
     @Override
@@ -75,10 +76,15 @@ public class UserServiceImpl implements UserService {
         User user = userRepository.findByEmail(email)
                 .orElseThrow(UserNotFoundException::new);
         userMapper.updateUserFromDto(dto, user);
-        // TODO update image logic
         userRepository.save(user);
         cacheService.invalidateUserById(user.getId());
         return userMapper.userToUserDto(user);
+    }
+
+    @Override
+    public void updatePhoto(MultipartFile file, String email) {
+        UUID id = userRepository.findIdByEmail(email);
+        // todo update image
     }
 
     @Override
